requirements-dev.txt
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0
pytest-mock==3.11.1 